fix(markers): address review findings on the marker store
Batches: the Kotlin header length is summed in Long, so counts whose byte
product overflows Int no longer pass the length check, and a corrupt batch
that still fails to decode is dropped instead of escaping the store thread.
Both stores refuse handles at or above 1 << 22 and any handle more than
65,536 past the current arrays, so one bad record cannot make them allocate
gigabytes.

Threads: the Kotlin viewport refresh holds the store lock only for the index
query and the final materialize, not through the cluster pass, and
`markerCount` is a volatile read, so a camera move on the main thread no
longer waits for background geometry. Both stores coalesce listener
notifications to one pending post, so a stream of position batches drives one
diff per burst instead of one per batch.

Geometry: the viewport filter on both platforms handles bounds that cross the
antimeridian (Android used to blank the whole layer there, iOS dropped the
far side), and both spatial indexes reject queries that miss the grid in
longitude instead of returning the outermost column.

JS: `set()` frees the handles of removed markers before assigning new ones, so
a full replacement reuses them; `<Marker>` child callbacks no longer fire for
a collection marker with the same id. The benchmark's collection motion now
matches the array motion, the demo drops stale cluster lookups, and the
JS-lag threshold names the scenarios it covers.

+++ package/android/src/main/java/com/margelo/nitro/nitromaps/MarkerBatchDecoder.kt
@@ -51,11 +51,12 @@ internal class MarkerBatchHeader(
   val removeCount: Int,
   val positionCount: Int,
 ) {
-  val totalBytes: Int
-    get() = MarkerBatchLayout.HEADER_BYTES +
-      upsertCount * MarkerBatchLayout.UPSERT_BYTES +
-      removeCount * MarkerBatchLayout.REMOVE_BYTES +
-      positionCount * MarkerBatchLayout.POSITION_BYTES
+  /** In `Long`: the counts are untrusted and their products overflow `Int`. */
+  val totalBytes: Long
+    get() = MarkerBatchLayout.HEADER_BYTES.toLong() +
+      upsertCount.toLong() * MarkerBatchLayout.UPSERT_BYTES +
+      removeCount.toLong() * MarkerBatchLayout.REMOVE_BYTES +
+      positionCount.toLong() * MarkerBatchLayout.POSITION_BYTES
 }
 
 class MalformedMarkerBatchException(message: String) : IllegalArgumentException(message)
@@ -82,7 +83,7 @@ internal object MarkerBatchDecoder {
       positionCount = buffer.getInt(12),
     )
     if (header.upsertCount < 0 || header.removeCount < 0 || header.positionCount < 0 ||
-      header.totalBytes != buffer.limit()
+      header.totalBytes != buffer.limit().toLong()
     ) {
       throw MalformedMarkerBatchException(
         "Marker batch is ${buffer.limit()} bytes, expected ${header.totalBytes}",
